[Feature] Added LocationsViewModel object vm to environment, Extracted LocationsViewModel into LocationsViewModel file from LocationsView file
