+++ SwiftfulMapApp/Views/LocationsView.swift
@@ -7,20 +7,11 @@
 
 import SwiftUI
 
-class LocationsViewModel: ObservableObject{
-    
-    @Published var locations: [Location]
-    
-    init() {
-        let locations = LocationsDataService.locations
-        self.locations = locations
-    }
-    
-}
+
 
 struct LocationsView: View {
     
-    @StateObject private var vm = LocationsViewModel()
+    @EnvironmentObject private var vm: LocationsViewModel
     
     var body: some View {
         List{
@@ -33,4 +24,5 @@ struct LocationsView: View {
 
 #Preview {
     LocationsView()
+        .environmentObject(LocationsViewModel())
 }
